streamlit_app/requirements.txt
openai
streamlit
torch
openai-whisper
google-generativeai
ffmpeg-python
imageio-ffmpeg
python-dotenv
audio-recorder-streamlit
streamlit-float
pyttsx3
yfinance
unstructured
beautifulsoup4 
requests
choco 
ffmpeg
scipy
